Extract map logic to its own file
